requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.3
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.0.1
jsonschema==3.2.0
jwcrypto==0.8
Markdown==3.3.4
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
pyrsistent==0.17.3
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
